<feature> : <Backend API 인가를 위한 파일 분리>
Backend API 사용에 있어서 권한 설정을 위해서 js 파일을 분리
분리한 내용에 맞게, import문 수정
Tester, Maker 각각만 사용하는 컴포넌트, 뷰 분리

Resolves: #issueNo
See also: #issueNo
