+++ src/views/maker/FixTestView.vue
@@ -91,13 +91,15 @@
 </template>
 
 <script>
-import { createTest } from '@/api/auth';
+import { findDetailTest } from '@/api/auth';
+import { updateTest } from '@/api/makerAuth';
 
 export default {
-  name: 'CreateTestView',
+  name: 'FixTestView',
   data() {
     return {
       valid: null,
+      id: '',
       title: '',
       recruitmentTimeStart: '',
       recruitmentTimeLimit: '',
@@ -130,7 +132,11 @@ export default {
         for (let key of testData.entries()) {
           console.log(`${key}`);
         }
-        const response = await createTest(this.$store.state.UserID, testData);
+        const response = await updateTest(
+          this.$store.state.UserID,
+          this.id,
+          testData,
+        );
         console.log(response.data);
 
         await this.$router.push('/main');
@@ -178,6 +184,24 @@ export default {
       }
       this.overlay = false;
     },
+    async initValue() {
+      const detailTest = await findDetailTest(this.$route.params.id);
+      const detailTestData = detailTest.data;
+      console.log(detailTestData);
+
+      this.id = detailTestData.id;
+      this.title = detailTestData.title;
+      this.recruitmentTimeStart = detailTestData.recruitmentTimeStart;
+      this.recruitmentTimeLimit = detailTestData.recruitmentTimeLimit;
+      this.durationTimeStart = detailTestData.durationTimeStart;
+      this.durationTimeLimit = detailTestData.durationTimeLimit;
+      this.participantCapacity = detailTestData.participantCapacity;
+      this.reward = detailTestData.reward;
+      this.content = detailTestData.content;
+    },
+  },
+  created() {
+    this.initValue();
   },
 };
 </script>
